Split test_parametric_ops
Move test cases exclusively for controlled parametric ops to
tests/ops/op_math/test_parametric_controlled_ops.py
