+++ tests/ops/op_math/test_parametric_controlled_ops.py
@@ -15,8 +15,11 @@
 Unit tests for parametric Operators inheriting from ControlledOp.
 """
 
+from functools import reduce
+
 import numpy as np
 import pytest
+from gate_data import ControlledPhaseShift
 from scipy.linalg import fractional_matrix_power
 from scipy.stats import unitary_group
 
@@ -26,6 +29,14 @@ from pennylane.ops.qubit.matrix_ops import QubitUnitary
 
 X = np.array([[0, 1], [1, 0]])
 X_broadcasted = np.array([X] * 3)
+
+
+def dot_broadcasted(a, b):
+    return np.einsum("...ij,...jk->...ik", a, b)
+
+
+def multi_dot_broadcasted(matrices):
+    return reduce(dot_broadcasted, matrices)
 
 
 class TestControlledQubitUnitary:
@@ -431,3 +442,633 @@ class TestControlledQubitUnitary:
             qml.ControlledQubitUnitary(
                 not_unitary, control_wires=[0, 2], wires=1, unitary_check=True
             )
+
+
+class TestDecompositions:
+    def test_CRX_decomposition(self):
+        """Test the decomposition for CRX."""
+        phi = 0.432
+
+        ops1 = qml.CRX.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRX(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.RZ, qml.RY, qml.CNOT, qml.RY, qml.CNOT, qml.RZ]
+        params = [[np.pi / 2], [phi / 2], [], [-phi / 2], [], [-np.pi / 2]]
+        wires = [Wires(1), Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1)), Wires(1)]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert op.parameters == p
+                assert op.wires == w
+
+    def test_CRX_decomposition_broadcasted(self):
+        """Test the decomposition for broadcasted CRX."""
+        phi = np.array([0.1, 2.1])
+
+        ops1 = qml.CRX.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRX(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.RZ, qml.RY, qml.CNOT, qml.RY, qml.CNOT, qml.RZ]
+        params = [[np.pi / 2], [phi / 2], [], [-phi / 2], [], [-np.pi / 2]]
+        wires = [Wires(1), Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1)), Wires(1)]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert qml.math.allclose(op.parameters, p)
+                assert op.wires == w
+
+    def test_CRY_decomposition(self):
+        """Test the decomposition for CRY."""
+        phi = 0.432
+
+        ops1 = qml.CRY.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRY(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.RY, qml.CNOT, qml.RY, qml.CNOT]
+        params = [[phi / 2], [], [-phi / 2], []]
+        wires = [Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1))]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert np.allclose(op.parameters, p)
+                assert op.wires == w
+
+    def test_CRY_decomposition_broadcasted(self):
+        """Test the decomposition for broadcastedCRY."""
+        phi = np.array([2.1, 0.2])
+
+        ops1 = qml.CRY.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRY(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.RY, qml.CNOT, qml.RY, qml.CNOT]
+        params = [[phi / 2], [], [-phi / 2], []]
+        wires = [Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1))]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert np.allclose(op.parameters, p)
+                assert op.wires == w
+
+    def test_CRZ_decomposition(self):
+        """Test the decomposition for CRZ."""
+        phi = 0.321
+
+        ops1 = qml.CRZ.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRZ(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.PhaseShift, qml.CNOT, qml.PhaseShift, qml.CNOT]
+        params = [[phi / 2], [], [-phi / 2], []]
+        wires = [Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1))]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert np.allclose(op.parameters, p)
+                assert op.wires == w
+
+    def test_CRZ_decomposition_broadcasted(self):
+        """Test the decomposition for broadcasted CRZ."""
+        phi = np.array([0.6, 2.1])
+
+        ops1 = qml.CRZ.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRZ(phi, wires=(0, 1)).decomposition()
+
+        classes = [qml.PhaseShift, qml.CNOT, qml.PhaseShift, qml.CNOT]
+        params = [[phi / 2], [], [-phi / 2], []]
+        wires = [Wires(1), Wires((0, 1)), Wires(1), Wires((0, 1))]
+
+        for ops in [ops1, ops2]:
+            for op, c, p, w in zip(ops, classes, params, wires):
+                assert isinstance(op, c)
+                assert np.allclose(op.parameters, p)
+                assert op.wires == w
+
+    @pytest.mark.parametrize("phi, theta, omega", [[0.5, 0.6, 0.7], [0.1, -0.4, 0.7], [-10, 5, -1]])
+    def test_CRot_decomposition(self, tol, phi, theta, omega):
+        """Tests that the decomposition of the CRot gate is correct"""
+        op = qml.CRot(phi, theta, omega, wires=[0, 1])
+        res = op.decomposition()
+
+        mats = []
+        for i in reversed(res):
+            if len(i.wires) == 1:
+                mats.append(np.kron(np.eye(2), i.matrix()))
+            else:
+                mats.append(i.matrix())
+
+        decomposed_matrix = np.linalg.multi_dot(mats)
+
+        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
+
+    @pytest.mark.parametrize(
+        "phi, theta, omega",
+        [
+            [np.array([0.1, 0.2]), np.array([-0.4, 2.19]), np.array([0.7, -0.7])],
+            [np.array([0.1, 0.2, 0.9]), -0.4, np.array([0.7, 0.0, -0.7])],
+        ],
+    )
+    def test_CRot_decomposition_broadcasted(self, tol, phi, theta, omega):
+        """Tests that the decomposition of the broadcasted CRot gate is correct"""
+        op = qml.CRot(phi, theta, omega, wires=[0, 1])
+        res = op.decomposition()
+
+        mats = []
+        for i in reversed(res):
+            mat = i.matrix()
+            if len(i.wires) == 1:
+                I = np.eye(2)[np.newaxis] if qml.math.ndim(mat) == 3 else np.eye(2)
+                mats.append(np.kron(I, mat))
+            else:
+                mats.append(mat)
+
+        decomposed_matrix = multi_dot_broadcasted(mats)
+
+        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
+
+    @pytest.mark.parametrize("phi", [-0.1, 0.2, 0.5])
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_decomp(self, phi, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation
+        calculates the correct decomposition"""
+        op = cphase_op(phi, wires=[0, 2])
+        decomp = op.decomposition()
+
+        mats = []
+        for i in reversed(decomp):
+            if i.wires.tolist() == [0]:
+                mats.append(np.kron(i.matrix(), np.eye(4)))
+            elif i.wires.tolist() == [1]:
+                mats.append(np.kron(np.eye(2), np.kron(i.matrix(), np.eye(2))))
+            elif i.wires.tolist() == [2]:
+                mats.append(np.kron(np.eye(4), i.matrix()))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 1]:
+                mats.append(np.kron(i.matrix(), np.eye(2)))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 2]:
+                mats.append(
+                    np.array(
+                        [
+                            [1, 0, 0, 0, 0, 0, 0, 0],
+                            [0, 1, 0, 0, 0, 0, 0, 0],
+                            [0, 0, 1, 0, 0, 0, 0, 0],
+                            [0, 0, 0, 1, 0, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 1, 0, 0],
+                            [0, 0, 0, 0, 1, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 0, 1],
+                            [0, 0, 0, 0, 0, 0, 1, 0],
+                        ]
+                    )
+                )
+
+        decomposed_matrix = np.linalg.multi_dot(mats)
+        lam = np.exp(1j * phi)
+        exp = np.array(
+            [
+                [1, 0, 0, 0, 0, 0, 0, 0],
+                [0, 1, 0, 0, 0, 0, 0, 0],
+                [0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 1, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0],
+                [0, 0, 0, 0, 0, lam, 0, 0],
+                [0, 0, 0, 0, 0, 0, 1, 0],
+                [0, 0, 0, 0, 0, 0, 0, lam],
+            ]
+        )
+
+        assert np.allclose(decomposed_matrix, exp)
+
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_decomp_broadcasted(self, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation
+        calculates the correct decomposition"""
+        phi = np.array([-0.2, 4.2, 1.8])
+        op = cphase_op(phi, wires=[0, 2])
+        decomp = op.decomposition()
+
+        mats = []
+        for i in reversed(decomp):
+            mat = i.matrix()
+            eye = np.eye(2)[np.newaxis] if np.ndim(mat) == 3 else np.eye(2)
+            if i.wires.tolist() == [0]:
+                mats.append(np.kron(mat, np.kron(eye, eye)))
+            elif i.wires.tolist() == [1]:
+                mats.append(np.kron(eye, np.kron(mat, eye)))
+            elif i.wires.tolist() == [2]:
+                mats.append(np.kron(np.kron(eye, eye), mat))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 1]:
+                mats.append(np.kron(mat, eye))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 2]:
+                mats.append(
+                    np.array(
+                        [
+                            [1, 0, 0, 0, 0, 0, 0, 0],
+                            [0, 1, 0, 0, 0, 0, 0, 0],
+                            [0, 0, 1, 0, 0, 0, 0, 0],
+                            [0, 0, 0, 1, 0, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 1, 0, 0],
+                            [0, 0, 0, 0, 1, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 0, 1],
+                            [0, 0, 0, 0, 0, 0, 1, 0],
+                        ]
+                    )
+                )
+
+        decomposed_matrix = multi_dot_broadcasted(mats)
+        lam = np.exp(1j * phi)
+        exp = np.array([np.diag([1, 1, 1, 1, 1, el, 1, el]) for el in lam])
+
+        assert np.allclose(decomposed_matrix, exp)
+
+
+class TestMatrix:
+    def test_CRX(self, tol):
+        """Test controlled x rotation is correct"""
+
+        # test identity for theta=0
+        assert np.allclose(qml.CRX.compute_matrix(0), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRX(0, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, 1 / np.sqrt(2), -1j / np.sqrt(2)],
+                [0, 0, -1j / np.sqrt(2), 1 / np.sqrt(2)],
+            ]
+        )
+        assert np.allclose(qml.CRX.compute_matrix(np.pi / 2), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRX(np.pi / 2, wires=[0, 1]).matrix(), expected_pi_half, atol=tol, rtol=0
+        )
+
+        # test identity for theta=pi
+        expected_pi = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1j], [0, 0, -1j, 0]])
+        assert np.allclose(qml.CRX.compute_matrix(np.pi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRX(np.pi, wires=[0, 1]).matrix(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        expected = [expected_pi_half, expected_pi]
+        assert np.allclose(qml.CRX.compute_matrix(param), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRX(param, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0)
+
+    def test_CRY(self, tol):
+        """Test controlled y rotation is correct"""
+
+        # test identity for theta=0
+        assert np.allclose(qml.CRY.compute_matrix(0), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRY(0, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, 1 / np.sqrt(2), -1 / np.sqrt(2)],
+                [0, 0, 1 / np.sqrt(2), 1 / np.sqrt(2)],
+            ]
+        )
+        assert np.allclose(qml.CRY.compute_matrix(np.pi / 2), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRY(np.pi / 2, wires=[0, 1]).matrix(), expected_pi_half, atol=tol, rtol=0
+        )
+
+        # test identity for theta=pi
+        expected_pi = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])
+        assert np.allclose(qml.CRY.compute_matrix(np.pi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRY(np.pi, wires=[0, 1]).matrix(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        expected = [expected_pi_half, expected_pi]
+        assert np.allclose(qml.CRY.compute_matrix(param), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRY(param, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0)
+
+    def test_CRZ(self, tol):
+        """Test controlled z rotation is correct"""
+
+        # test identity for theta=0
+        assert np.allclose(qml.CRZ.compute_matrix(0), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(0, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, np.exp(-1j * np.pi / 4), 0],
+                [0, 0, 0, np.exp(1j * np.pi / 4)],
+            ]
+        )
+        assert np.allclose(qml.CRZ.compute_matrix(np.pi / 2), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRZ(np.pi / 2, wires=[0, 1]).matrix(), expected_pi_half, atol=tol, rtol=0
+        )
+
+        # test identity for theta=pi
+        expected_pi = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, -1j, 0], [0, 0, 0, 1j]])
+        assert np.allclose(qml.CRZ.compute_matrix(np.pi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(np.pi, wires=[0, 1]).matrix(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        expected = [expected_pi_half, expected_pi]
+        assert np.allclose(qml.CRZ.compute_matrix(param), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(param, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    def test_CRZ_tf(self, tol):
+        """Test controlled z rotation is correct when used with Tensorflow,
+        because the code differs in that case."""
+        import tensorflow as tf
+
+        # test identity for theta=0
+        z = tf.Variable(0.0)
+        assert np.allclose(qml.CRZ.compute_matrix(z), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(z, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, np.exp(-1j * np.pi / 4), 0],
+                [0, 0, 0, np.exp(1j * np.pi / 4)],
+            ]
+        )
+        phi = tf.Variable(np.pi / 2)
+        assert np.allclose(qml.CRZ.compute_matrix(phi), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(phi, wires=[0, 1]).matrix(), expected_pi_half, atol=tol, rtol=0)
+
+        # test identity for theta=pi
+        expected_pi = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, -1j, 0], [0, 0, 0, 1j]])
+        phi = tf.Variable(np.pi)
+        assert np.allclose(qml.CRZ.compute_matrix(phi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(phi, wires=[0, 1]).matrix(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        expected = [expected_pi_half, expected_pi]
+        param_tf = tf.Variable(param)
+        assert np.allclose(qml.CRZ.compute_matrix(param_tf), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(param_tf, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0)
+
+    def test_CRot(self, tol):
+        """Test controlled arbitrary rotation is correct"""
+
+        # test identity for phi,theta,omega=0
+        assert np.allclose(qml.CRot.compute_matrix(0, 0, 0), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRot(0, 0, 0, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0
+        )
+
+        # test identity for phi,theta,omega=pi
+        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])
+        assert np.allclose(qml.CRot.compute_matrix(np.pi, np.pi, np.pi), expected, atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRot(np.pi, np.pi, np.pi, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0
+        )
+
+        def arbitrary_Crotation(x, y, z):
+            """controlled arbitrary single qubit rotation"""
+            c = np.cos(y / 2)
+            s = np.sin(y / 2)
+            return np.array(
+                [
+                    [1, 0, 0, 0],
+                    [0, 1, 0, 0],
+                    [0, 0, np.exp(-0.5j * (x + z)) * c, -np.exp(0.5j * (x - z)) * s],
+                    [0, 0, np.exp(-0.5j * (x - z)) * s, np.exp(0.5j * (x + z)) * c],
+                ]
+            )
+
+        a, b, c = 0.432, -0.152, 0.9234
+        assert np.allclose(
+            qml.CRot.compute_matrix(a, b, c), arbitrary_Crotation(a, b, c), atol=tol, rtol=0
+        )
+        assert np.allclose(
+            qml.CRot(a, b, c, wires=[0, 1]).matrix(),
+            arbitrary_Crotation(a, b, c),
+            atol=tol,
+            rtol=0,
+        )
+
+    def test_CRot_broadcasted(self, tol):
+        """Test broadcasted controlled arbitrary rotation is correct"""
+
+        # test identity for phi,theta,omega=0
+        z = np.zeros(5)
+        assert np.allclose(qml.CRot.compute_matrix(z, z, z), np.identity(4), atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRot(z, z, z, wires=[0, 1]).matrix(), np.identity(4), atol=tol, rtol=0
+        )
+
+        # test -i*CY for phi,theta,omega=pi
+        expected = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])
+        pi = np.ones(3) * np.pi
+        assert np.allclose(qml.CRot.compute_matrix(pi, pi, pi), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRot(pi, pi, pi, wires=[0, 1]).matrix(), expected, atol=tol, rtol=0)
+
+        def arbitrary_Crotation(x, y, z):
+            """controlled arbitrary single qubit rotation"""
+            c = np.cos(y / 2)
+            s = np.sin(y / 2)
+            return np.array(
+                [
+                    [
+                        [1, 0, 0, 0],
+                        [0, 1, 0, 0],
+                        [0, 0, np.exp(-0.5j * (_x + _z)) * _c, -np.exp(0.5j * (_x - _z)) * _s],
+                        [0, 0, np.exp(-0.5j * (_x - _z)) * _s, np.exp(0.5j * (_x + _z)) * _c],
+                    ]
+                    for _x, _z, _c, _s in zip(x, z, c, s)
+                ]
+            )
+
+        a, b, c = np.array([0.432, -0.124]), np.array([-0.152, 2.912]), np.array([0.9234, -9.2])
+        assert np.allclose(
+            qml.CRot.compute_matrix(a, b, c), arbitrary_Crotation(a, b, c), atol=tol, rtol=0
+        )
+        assert np.allclose(
+            qml.CRot(a, b, c, wires=[0, 1]).matrix(),
+            arbitrary_Crotation(a, b, c),
+            atol=tol,
+            rtol=0,
+        )
+
+    @pytest.mark.parametrize("phi", [-0.1, 0.2, 0.5])
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_matrix_and_eigvals(self, phi, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation calculate the correct
+        matrix and eigenvalues"""
+        op = cphase_op(phi, wires=[0, 1])
+        res = op.matrix()
+        exp = ControlledPhaseShift(phi)
+        assert np.allclose(res, exp)
+
+        res = op.eigvals()
+        assert np.allclose(np.diag(res), exp)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("phi", [-0.1, 0.2, 0.5])
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_matrix_and_eigvals_tf(self, phi, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation calculate the correct
+        matrix and eigenvalues for the Tensorflow interface, because the code differs
+        in that case."""
+        import tensorflow as tf
+
+        op = cphase_op(tf.Variable(phi), wires=[0, 1])
+        res = op.matrix()
+        exp = ControlledPhaseShift(phi)
+        assert np.allclose(res, exp)
+
+        res = op.eigvals()
+        assert np.allclose(np.diag(res), exp)
+
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_matrix_and_eigvals_broadcasted(self, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation calculate the
+        correct matrix and eigenvalues for broadcasted parameters"""
+        phi = np.array([0.2, np.pi / 2, -0.1])
+        op = cphase_op(phi, wires=[0, 1])
+        res = op.matrix()
+        expected = np.array([np.eye(4, dtype=complex)] * 3)
+        expected[..., 3, 3] = np.exp(1j * phi)
+        assert np.allclose(res, expected)
+
+        res = op.eigvals()
+        exp_eigvals = np.ones((3, 4), dtype=complex)
+        exp_eigvals[..., 3] = np.exp(1j * phi)
+        assert np.allclose(res, exp_eigvals)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("cphase_op", [qml.ControlledPhaseShift, qml.CPhase])
+    def test_controlled_phase_shift_matrix_and_eigvals_broadcasted_tf(self, cphase_op):
+        """Tests that the ControlledPhaseShift and CPhase operation calculate the
+        correct matrix and eigenvalues for broadcasted parameters and Tensorflow,
+        because the code differs for that interface."""
+        import tensorflow as tf
+
+        phi = np.array([0.2, np.pi / 2, -0.1])
+        phi_tf = tf.Variable(phi)
+        op = cphase_op(phi_tf, wires=[0, 1])
+        res = op.matrix()
+        expected = np.array([np.eye(4, dtype=complex)] * 3)
+        expected[..., 3, 3] = np.exp(1j * phi)
+        assert np.allclose(res, expected)
+
+        res = op.eigvals()
+        exp_eigvals = np.ones((3, 4), dtype=complex)
+        exp_eigvals[..., 3] = np.exp(1j * phi)
+        assert np.allclose(res, exp_eigvals)
+
+
+class TestEigvals:  # pylint: disable=too-few-public-methods
+    """Tests for the Eigvals operation"""
+
+    def test_crz_eigvals(self, tol):
+        """Test controlled z rotation eigvals are correct"""
+
+        # test identity for theta=0
+        assert np.allclose(qml.CRZ.compute_eigvals(0), np.ones(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(0, wires=[0, 1]).eigvals(), np.ones(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array([1, 1, np.exp(-1j * np.pi / 4), np.exp(1j * np.pi / 4)])
+        assert np.allclose(qml.CRZ.compute_eigvals(np.pi / 2), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(
+            qml.CRZ(np.pi / 2, wires=[0, 1]).eigvals(), expected_pi_half, atol=tol, rtol=0
+        )
+
+        # test identity for theta=pi
+        expected_pi = np.array([1, 1, -1j, 1j])
+        assert np.allclose(qml.CRZ.compute_eigvals(np.pi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(np.pi, wires=[0, 1]).eigvals(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        expected = [expected_pi_half, expected_pi]
+        assert np.allclose(qml.CRZ.compute_eigvals(param), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(param, wires=[0, 1]).eigvals(), expected, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    def test_crz_eigvals_tf(self, tol):
+        """Test controlled z rotation eigvals are correct with Tensorflow, because the
+        code differs for that interface."""
+        import tensorflow as tf
+
+        # test identity for theta=0
+        z = tf.Variable(0)
+        assert np.allclose(qml.CRZ.compute_eigvals(z), np.ones(4), atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(z, wires=[0, 1]).eigvals(), np.ones(4), atol=tol, rtol=0)
+
+        # test identity for theta=pi/2
+        expected_pi_half = np.array([1, 1, np.exp(-1j * np.pi / 4), np.exp(1j * np.pi / 4)])
+        phi = tf.Variable(np.pi / 2)
+        assert np.allclose(qml.CRZ.compute_eigvals(phi), expected_pi_half, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(phi, wires=[0, 1]).eigvals(), expected_pi_half, atol=tol, rtol=0)
+
+        # test identity for theta=pi
+        expected_pi = np.array([1, 1, -1j, 1j])
+        phi = tf.Variable(np.pi)
+        assert np.allclose(qml.CRZ.compute_eigvals(phi), expected_pi, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(phi, wires=[0, 1]).eigvals(), expected_pi, atol=tol, rtol=0)
+
+        # test broadcasting
+        param = np.array([np.pi / 2, np.pi])
+        param_tf = tf.Variable(param)
+        expected = [expected_pi_half, expected_pi]
+        assert np.allclose(qml.CRZ.compute_eigvals(param_tf), expected, atol=tol, rtol=0)
+        assert np.allclose(qml.CRZ(param_tf, wires=[0, 1]).eigvals(), expected, atol=tol, rtol=0)
+
+
+def test_simplify_crot():
+    """Simplify CRot operations with different parameters."""
+
+    crot_x = qml.CRot(np.pi / 2, 0.1, -np.pi / 2, wires=[0, 1])
+    simplify_crot_x = crot_x.simplify()
+
+    assert simplify_crot_x.name == "CRX"
+    assert simplify_crot_x.data == (0.1,)
+    assert np.allclose(simplify_crot_x.matrix(), crot_x.matrix())
+
+    crot_y = qml.CRot(0, 0.1, 0, wires=[0, 1])
+    simplify_crot_y = crot_y.simplify()
+
+    assert simplify_crot_y.name == "CRY"
+    assert simplify_crot_y.data == (0.1,)
+    assert np.allclose(simplify_crot_y.matrix(), crot_y.matrix())
+
+    crot_z = qml.CRot(0.1, 0, 0.2, wires=[0, 1])
+    simplify_crot_z = crot_z.simplify()
+
+    assert simplify_crot_z.name == "CRZ"
+    assert np.allclose(simplify_crot_z.data, [0.3])
+    assert np.allclose(simplify_crot_z.matrix(), crot_z.matrix())
+
+    crot = qml.CRot(0.1, 0.2, 0.3, wires=[0, 1])
+    not_simplified_crot = crot.simplify()
+
+    assert not_simplified_crot.name == "CRot"
+    assert np.allclose(not_simplified_crot.matrix(), crot.matrix())
+
+
+controlled_data = [
+    (qml.RX(1.234, wires=0), qml.CRX(1.234, wires=("a", 0))),
+    (qml.RY(1.234, wires=0), qml.CRY(1.234, wires=("a", 0))),
+    (qml.RZ(1.234, wires=0), qml.CRZ(1.234, wires=("a", 0))),
+    (qml.PhaseShift(1.234, wires=0), qml.ControlledPhaseShift(1.234, wires=("a", 0))),
+    (qml.Rot(1.2, 2.3, 3.4, wires=0), qml.CRot(1.2, 2.3, 3.4, wires=("a", 0))),
+]
+
+
+@pytest.mark.parametrize("base, cbase", controlled_data)
+def test_controlled_method(base, cbase):
+    """Tests the _controlled method for parametric ops."""
+    # pylint: disable=protected-access
+    assert qml.equal(base._controlled("a"), cbase)
